Refactor & bug fixes
