+++ build.gradle
@@ -32,6 +32,7 @@ dependencies {
     compile('org.springframework.boot:spring-boot-starter-freemarker')
     compile("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
     compile("org.jetbrains.kotlin:kotlin-reflect")
+    compile('com.fasterxml.jackson.module:jackson-module-kotlin')
     testCompile('org.springframework.boot:spring-boot-starter-test')
 }
