+++ src/main/kotlin/ru/kjudge/common/entity/Limits.kt
@@ -1,4 +1,4 @@
-package ru.kjudge.web.entity
+package ru.kjudge.common.entity
 
 data class Limits(
         /**
