+++ src/main/kotlin/ru/kjudge/common/entity/Test.kt
@@ -1,4 +1,4 @@
-package ru.kjudge.web.entity
+package ru.kjudge.common.entity
 
 data class Test(
         val test: String,
