+++ src/main/kotlin/ru/kjudge/web/RunnerClient.kt
@@ -4,10 +4,12 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.amqp.core.DirectExchange
+import org.springframework.amqp.rabbit.AsyncRabbitTemplate
 import org.springframework.amqp.rabbit.core.RabbitTemplate
+import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
-import ru.kjudge.web.entity.Message
-import ru.kjudge.web.entity.RunResult
+import ru.kjudge.common.entity.Message
+import ru.kjudge.common.entity.RunResult
 
 
 @Service
@@ -15,14 +17,18 @@ class RunnerClient {
     private val log: Logger = LoggerFactory.getLogger(RunnerClient::class.java.name)
 
     @Autowired
-    lateinit var template: RabbitTemplate
+    lateinit var template: AsyncRabbitTemplate
 
     @Autowired
     lateinit var exchange: DirectExchange
 
     fun send(message: Message) {
-        val result = template.convertSendAndReceive(exchange.name, "rpc", message) as RunResult
-        log.info("Result from running server: $result")
+        val promise = template.convertSendAndReceive<RunResult>(exchange.name, "rpc", message)
+        promise.addCallback({
+            log.info("Result from running server: $it")
+        }, {
+            log.info("Error!!!!1!")
+        })
     }
 
 }
